Fix (pdfwrite) : (1) For better precision changing the internal representation of Widths to double, and skipping small noise while writing Tj. (2) Widths for Type 3 fonts were not computed (020-01.ps, 031-01.ps).
git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@3038 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/src/gdevpdtf.h
@@ -158,7 +158,7 @@ struct pdf_font_resource_s {
     pdf_base_font_t *base_font;	/* == FontDescriptor->base_font */
     gs_font_base *copied_font;	/* == base_font->copied */
     uint count;			/* # of chars/CIDs */
-    int *Widths;		/* [count] (not used for Type 0) */
+    double *Widths;		/* [count] (not used for Type 0) */
     byte *used;			/* [ceil(count/8)] bitmap of chars/CIDs used */
 				/* (not used for Type 0 or Type 3) */
     pdf_resource_t *ToUnicode;	/* CMap (not used for CIDFonts) */
